Unify function for creating shard key, we had it in two places (#6227)
